# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Payroll Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalEmployees: number;
  activeEmployees: number;
  totalPayrollThisMonth: number;
  pendingPayslips: number;
  /** @nullable */
  currentPeriodStatus: string | null;
  totalDepartments: number;
}