# lib/api-zod/src/generated/types/departmentUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Payroll Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface DepartmentUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
}