# lib/api-zod/src/generated/types/employee.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Payroll Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { EmployeeStatus } from './employeeStatus';

export interface Employee {
  id: number;
  employeeCode: string;
  name: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  position: string;
  departmentId: number;
  /** @nullable */
  departmentName?: string | null;
  baseSalary: number;
  transportAllowance?: number;
  mealAllowance?: number;
  status: EmployeeStatus;
  joinDate: string;
  createdAt: string;
}