# lib/api-zod/src/generated/types/employeeInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Payroll Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { EmployeeInputStatus } from './employeeInputStatus';

export interface EmployeeInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  employeeCode: string;
  email?: string;
  phone?: string;
  /** @minLength 1 */
  position: string;
  departmentId: number;
  /** @minimum 0 */
  baseSalary: number;
  /** @minimum 0 */
  transportAllowance?: number;
  /** @minimum 0 */
  mealAllowance?: number;
  status?: EmployeeInputStatus;
  joinDate: string;
}