# lib/api-zod/src/generated/types/payrollPeriod.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Payroll Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { PayrollPeriodStatus } from './payrollPeriodStatus';

export interface PayrollPeriod {
  id: number;
  /**
     * @minimum 1
     * @maximum 12
     */
  month: number;
  year: number;
  status: PayrollPeriodStatus;
  /** @nullable */
  totalPayroll?: number | null;
  /** @nullable */
  totalEmployees?: number | null;
  /** @nullable */
  processedAt?: string | null;
  /** @nullable */
  paidAt?: string | null;
  createdAt: string;
}