# lib/api-zod/src/generated/types/payrollPeriodInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Payroll Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface PayrollPeriodInput {
  /**
     * @minimum 1
     * @maximum 12
     */
  month: number;
  year: number;
}