# lib/api-zod/src/generated/types/payrollPeriodUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Payroll Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { PayrollPeriodUpdateStatus } from './payrollPeriodUpdateStatus';

export interface PayrollPeriodUpdate {
  status?: PayrollPeriodUpdateStatus;
}