# lib/api-zod/src/generated/types/payslip.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Payroll Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { PayslipStatus } from './payslipStatus';

export interface Payslip {
  id: number;
  employeeId: number;
  /** @nullable */
  employeeName?: string | null;
  /** @nullable */
  employeeCode?: string | null;
  /** @nullable */
  position?: string | null;
  /** @nullable */
  departmentName?: string | null;
  periodId: number;
  /** @nullable */
  periodMonth?: number | null;
  /** @nullable */
  periodYear?: number | null;
  baseSalary: number;
  transportAllowance: number;
  mealAllowance: number;
  grossSalary: number;
  bpjsKetenagakerjaan: number;
  bpjsKesehatan: number;
  incomeTax: number;
  totalDeductions: number;
  netSalary: number;
  status: PayslipStatus;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}